refactor: 회원가입 UI 리펙토링
