+++ src/app/(login)/signup/(path)/email/page.tsx
@@ -1,32 +1,29 @@
 "use client";
 
-import React from "react";
 import { Controller } from "react-hook-form";
 
-import NavHeader from "@/components/NavHeader";
-import InputField, { InputFieldButton } from "@/components/base/InputField";
-import Label from "@/components/base/Label";
+import PageLayout from "@/app/(login)/signin/@components/PageLayout";
+import NavFooter from "@/components/NavFooter";
+import SHInputField, { InputFieldButton } from "@/components/base/SHInputField";
+import SHLabel from "@/components/base/SHLabel";
 import VStack from "@/components/base/stack/VStack";
 
-import ConfirmButton from "../../@components/ConfirmButton";
 import useSignUp from "../../@hooks/useSignUp";
 
 export default function page() {
-  const { control, handleConfirmEmail } = useSignUp();
+  const { router, control, handleConfirmEmail } = useSignUp();
 
   return (
-    <VStack>
-      <NavHeader />
-
-      <VStack className="mt-[16px] px-[20px]">
-        <Label type="SubTitle1">이메일을 인증해주세요.</Label>
+    <PageLayout className="bg-white">
+      <VStack>
+        <SHLabel type="SubTitle1">이메일을 인증해주세요.</SHLabel>
         <VStack className="mt-[50px] gap-[50px]">
           <Controller
             name="email"
             defaultValue={""}
             control={control}
             render={({ field }) => (
-              <InputField
+              <SHInputField
                 {...field}
                 label="이메일 주소"
                 type="email"
@@ -41,12 +38,18 @@ export default function page() {
             defaultValue={undefined}
             control={control}
             render={({ field }) => (
-              <InputField {...field} type="text" autoComplete="one-time-code" label="인증번호" />
+              <SHInputField {...field} type="text" autoComplete="one-time-code" label="인증번호" />
             )}
           />
-          <ConfirmButton onClick={handleConfirmEmail}>확인</ConfirmButton>
         </VStack>
       </VStack>
-    </VStack>
+      <VStack className="mt-[50px]">
+        <NavFooter
+          ratio="1:3"
+          left={{ onClick: router.back }}
+          right={{ onClick: handleConfirmEmail }}
+        />
+      </VStack>
+    </PageLayout>
   );
 }
