+++ src/app/(login)/signup/@hooks/useSignUp.tsx
@@ -6,7 +6,7 @@ import { useForm } from "react-hook-form";
 
 import { zodResolver } from "@hookform/resolvers/zod";
 
-import { SignUpScheme, SignUpType } from "../../signin/types/scheme";
+import { SignUpScheme, SignUpType } from "../../signin/@types/scheme";
 
 export default function useSignUp() {
   const router = useRouter();
@@ -26,6 +26,7 @@ export default function useSignUp() {
   };
 
   return {
+    router,
     control,
     handleConfirmEmail,
     handleSignUp,
